requirements.txt
altgraph==0.17
appdirs==1.4.4
arabic-reshaper==2.1.3
arduino-python3==0.6
asgiref==3.3.4
beautifulsoup4==4.9.1
click==7.1.2
cx-Freeze==6.1
distlib==0.3.1
Django==3.1.6
django-admin-bootstrapped==2.5.7
django-admin-interface==0.15.3  
django-appconf==1.0.4
django-autoslug==1.9.8
django-bootstrap-admin==0.3.4   
django-bootstrap-v5==1.0.3      
django-bootstrap4==3.0.1
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-cookies-samesite==0.9.0
django-crispy-forms==1.11.0
django-db==0.0.7
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-imagekit==4.0.2
django-js-asset==1.2.2
django-sb-admin==5.5.1.1
doqu==0.28.2
env==0.1.0
filelock==3.0.12
Flask==1.1.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
flaskwebgui==0.2.3
future==0.18.2
html5lib==1.1
importlib-metadata==2.1.1
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==2.0.1
nose==1.3.7
numpy==1.18.2
pandas==1.0.3
pefile==2019.4.18
pilkit==2.0
Pillow==8.1.0
psutil==5.8.0
pycparser==2.20
pyFirmata==1.1.0
PyInstaller==3.6
PyMySQL==1.0.2
PyPDF2==1.26.0
pyserial==3.4
python-bidi==0.4.2
python-dateutil==2.8.1
pythonnet==2.5.2
pytz==2019.3
pywebview==3.4
pywin32-ctypes==0.2.0
PyYAML==5.3.1
reportlab==3.5.67
selenium==3.141.0
serial==0.0.97
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.9
virtualenv==20.4.2
waitress==2.0.0
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
zipp==3.4.1